// src/app/features/pnl-and-funds-query/models/pnl-and-funds-query-response.model.ts
export interface PnlAndFundsQueryResponse {

    total_profit_tdy: string;
    unreal_tdy: string;
    profit_tdy: string;
    total_profit_amass: string;
    unreal_amass: string;
    profit_amass: string;
    avacapital: string;
    capital: string;
    hiscapital: string;
    tnetamt: string;
    t1netamt: string;
    t2netamt: string;
    tday: string;
    t1day: string;
    t2day: string;
    Currency: string;
    Currnm: string;
}